test(server): split report PDF test layers
Refs #2854
