+++ apps/server/tests/adapters/pdf/test_diagram_layout.py
@@ -102,9 +102,11 @@ def test_resolve_marker_states_classifies_correctly() -> None:
 _COLORS = {
     "brand": "#7c3aed",
     "axis": "#7b8da0",
+    "danger": "#c5221f",
     "text_secondary": "#52555e",
     "surface_alt": "#f1f2f6",
     "ink": "#1a1c24",
+    "success": "#0f9d58",
     "text_muted": "#6b6e78",
 }
 
@@ -147,6 +149,212 @@ def test_build_sensor_render_plan_multiple_sensors() -> None:
     assert markers[0].fill != markers[1].fill
     assert markers[0].outer_radius > markers[0].mid_radius > markers[0].radius
     assert markers[1].outer_radius > markers[1].mid_radius > markers[1].radius
+
+
+def test_sensor_state_mapping_connected_active_inactive_and_disconnected() -> None:
+    states = resolve_marker_states(
+        ["front-left wheel", "front-right wheel", "engine bay"],
+        connected_locations={"front-left wheel", "front-right wheel"},
+        amp_by_location={"front-left wheel": 21.0},
+    )
+    assert states["front-left wheel"] == "connected-active"
+    assert states["front-right wheel"] == "connected-inactive"
+    assert states["engine bay"] == "disconnected"
+
+
+def test_marker_gradients_keep_diagnostic_stroke_and_expand_with_intensity() -> None:
+    location_points_map = {
+        "front-left wheel": (40.0, 180.0),
+        "front-right wheel": (160.0, 180.0),
+        "rear-left wheel": (40.0, 80.0),
+        "engine bay": (100.0, 140.0),
+    }
+    markers, _, _ = build_sensor_render_plan(
+        location_points=location_points_map,
+        drawing_width=220.0,
+        drawing_height=252.0,
+        connected_locations={"front-left wheel", "front-right wheel", "rear-left wheel"},
+        amp_by_location={
+            "front-left wheel": 30.0,
+            "front-right wheel": 28.0,
+            "rear-left wheel": 15.0,
+        },
+        highlight={"rear-left wheel": _COLORS["danger"]},
+        colors=_COLORS,
+    )
+    marker_by_name = {marker.name: marker for marker in markers}
+
+    assert marker_by_name["rear-left wheel"].fill != _COLORS["danger"]
+    assert marker_by_name["rear-left wheel"].stroke == _COLORS["danger"]
+    assert (
+        marker_by_name["rear-left wheel"].outer_radius
+        > marker_by_name["rear-left wheel"].mid_radius
+    )
+    assert marker_by_name["rear-left wheel"].mid_radius > marker_by_name["rear-left wheel"].radius
+    assert marker_by_name["rear-left wheel"].outer_fill != marker_by_name["rear-left wheel"].fill
+    assert (
+        marker_by_name["front-left wheel"].outer_radius
+        > marker_by_name["front-left wheel"].mid_radius
+    )
+    assert marker_by_name["front-left wheel"].mid_radius > marker_by_name["front-left wheel"].radius
+    assert marker_by_name["front-left wheel"].stroke == marker_by_name["front-left wheel"].fill
+    assert marker_by_name["front-left wheel"].fill != marker_by_name["front-right wheel"].fill
+    assert marker_by_name["front-left wheel"].radius > marker_by_name["rear-left wheel"].radius
+    assert marker_by_name["engine bay"].state == "disconnected"
+    assert marker_by_name["engine bay"].outer_fill != marker_by_name["engine bay"].fill
+
+
+def test_marker_radius_grows_with_local_intensity_for_active_locations() -> None:
+    location_points_map = {
+        "front-left wheel": (40.0, 180.0),
+        "front-right wheel": (160.0, 180.0),
+        "rear-left wheel": (40.0, 80.0),
+    }
+    markers, _, _ = build_sensor_render_plan(
+        location_points=location_points_map,
+        drawing_width=220.0,
+        drawing_height=252.0,
+        connected_locations=set(location_points_map.keys()),
+        amp_by_location={
+            "front-left wheel": 12.0,
+            "front-right wheel": 22.0,
+            "rear-left wheel": 32.0,
+        },
+        highlight={},
+        colors=_COLORS,
+    )
+    marker_by_name = {marker.name: marker for marker in markers}
+
+    assert marker_by_name["front-left wheel"].radius < marker_by_name["front-right wheel"].radius
+    assert marker_by_name["front-right wheel"].radius < marker_by_name["rear-left wheel"].radius
+
+
+def test_highlighted_location_without_intensity_keeps_diagnostic_color() -> None:
+    location_points_map = {
+        "front-left wheel": (40.0, 180.0),
+        "front-right wheel": (160.0, 180.0),
+    }
+    markers, _, _ = build_sensor_render_plan(
+        location_points=location_points_map,
+        drawing_width=220.0,
+        drawing_height=252.0,
+        connected_locations=set(location_points_map.keys()),
+        amp_by_location={"front-right wheel": 22.0},
+        highlight={"front-left wheel": _COLORS["danger"]},
+        colors=_COLORS,
+    )
+    marker_by_name = {marker.name: marker for marker in markers}
+
+    assert marker_by_name["front-left wheel"].state == "connected-inactive"
+    assert marker_by_name["front-left wheel"].fill == _COLORS["danger"]
+    assert marker_by_name["front-left wheel"].stroke == _COLORS["danger"]
+
+
+def test_dense_layout_uses_gradient_markers_without_sensor_labels() -> None:
+    location_points_map = {
+        "front-left wheel": (36.0, 198.0),
+        "front-right wheel": (164.0, 198.0),
+        "rear-left wheel": (36.0, 76.0),
+        "rear-right wheel": (164.0, 76.0),
+        "engine bay": (100.0, 160.0),
+        "driveshaft tunnel": (100.0, 130.0),
+        "driver seat": (78.0, 145.0),
+        "trunk": (100.0, 102.0),
+    }
+    markers, labels, _ = build_sensor_render_plan(
+        location_points=location_points_map,
+        drawing_width=200.0,
+        drawing_height=252.0,
+        connected_locations=set(location_points_map.keys()),
+        amp_by_location={name: 10.0 + idx for idx, name in enumerate(location_points_map.keys())},
+        highlight={"front-left wheel": _COLORS["success"]},
+        colors=_COLORS,
+    )
+
+    assert labels == []
+    assert len(markers) == len(location_points_map)
+    for marker in markers:
+        assert marker.outer_radius > marker.mid_radius > marker.radius
+        assert marker.x - marker.outer_radius >= 0.0
+        assert marker.y - marker.outer_radius >= 0.0
+        assert marker.x + marker.outer_radius <= 200.0
+        assert marker.y + marker.outer_radius <= 252.0
+
+
+def test_single_sensor_uses_diagnostic_highlight_color() -> None:
+    markers, _, single_sensor = build_sensor_render_plan(
+        location_points={"front-left wheel": (40.0, 180.0)},
+        drawing_width=220.0,
+        drawing_height=252.0,
+        connected_locations={"front-left wheel"},
+        amp_by_location={"front-left wheel": 25.0},
+        highlight={"front-left wheel": _COLORS["danger"]},
+        colors=_COLORS,
+    )
+
+    assert single_sensor is True
+    assert markers[0].fill == _COLORS["danger"]
+    assert markers[0].stroke == markers[0].fill
+
+
+def test_narrow_page1_layout_returns_no_sensor_labels() -> None:
+    _, labels, _ = build_sensor_render_plan(
+        location_points={
+            "front-left wheel": (18.0, 198.0),
+            "front-right wheel": (106.0, 198.0),
+            "rear-left wheel": (18.0, 76.0),
+            "rear-right wheel": (106.0, 76.0),
+        },
+        drawing_width=124.0,
+        drawing_height=252.0,
+        connected_locations={
+            "front-left wheel",
+            "front-right wheel",
+            "rear-left wheel",
+            "rear-right wheel",
+        },
+        amp_by_location={
+            "front-left wheel": 31.1,
+            "front-right wheel": 34.9,
+            "rear-left wheel": 30.9,
+            "rear-right wheel": 31.4,
+        },
+        highlight={"rear-left wheel": _COLORS["danger"]},
+        colors=_COLORS,
+    )
+    assert labels == []
+
+
+def test_render_plan_prefers_diagnosed_location_when_intensity_hotspot_differs() -> None:
+    markers, _, _ = build_sensor_render_plan(
+        location_points={
+            "front-left wheel": (40.0, 180.0),
+            "front-right wheel": (160.0, 180.0),
+            "rear-left wheel": (40.0, 80.0),
+            "rear-right wheel": (160.0, 80.0),
+        },
+        drawing_width=220.0,
+        drawing_height=252.0,
+        connected_locations={
+            "front-left wheel",
+            "front-right wheel",
+            "rear-left wheel",
+            "rear-right wheel",
+        },
+        amp_by_location={
+            "front-left wheel": 36.0,
+            "front-right wheel": 34.0,
+            "rear-left wheel": 27.0,
+            "rear-right wheel": 26.0,
+        },
+        highlight={"rear-left wheel": _COLORS["danger"]},
+        colors=_COLORS,
+    )
+    marker_by_name = {marker.name: marker for marker in markers}
+
+    assert marker_by_name["front-left wheel"].fill != marker_by_name["rear-left wheel"].fill
+    assert marker_by_name["rear-left wheel"].fill != _COLORS["danger"]
+    assert marker_by_name["rear-left wheel"].stroke == _COLORS["danger"]
 
 
 # ── canonical_location ───────────────────────────────────────────────────────
