+++ apps/server/tests/adapters/pdf/test_report_pdf_content_rendering.py
@@ -1,0 +1,210 @@
+"""Rendered PDF content tests for headings, labels, and appendix wording."""
+
+from __future__ import annotations
+
+import json
+from pathlib import Path
+
+import pytest
+from _paths import SERVER_ROOT
+from test_support.core import extract_pdf_text
+from test_support.report_helpers import RUN_END, write_jsonl
+from test_support.report_helpers import report_run_metadata as _run_metadata
+from test_support.report_helpers import report_sample as _base_sample
+
+from vibesensor.adapters.analysis_summary import summarize_log
+from vibesensor.adapters.pdf.pdf_engine import build_report_pdf
+from vibesensor.shared.boundaries.reporting import prepare_report_input
+from vibesensor.shared.boundaries.reporting.document import (
+    AppendixAData,
+    AppendixCData,
+    MeasurementRow,
+    NextStep,
+    PatternEvidence,
+    RankedCandidateRow,
+    ReportDocument,
+    ReportLabelValueRow,
+    VerdictPageData,
+)
+from vibesensor.shared.constants.units import KMH_TO_MPS
+from vibesensor.use_cases.history.report_document import build_report_document
+
+_I18N_JSON = SERVER_ROOT / "data" / "report_i18n.json"
+
+
+def _sample(
+    idx: int,
+    *,
+    speed_kmh: float,
+    dominant_freq_hz: float,
+    peak_amp_g: float,
+) -> dict[str, object]:
+    return _base_sample(
+        idx,
+        speed_kmh=speed_kmh,
+        dominant_freq_hz=dominant_freq_hz,
+        peak_amp_g=peak_amp_g,
+        add_index_accel_offset=True,
+    )
+
+
+def _make_run_jsonl(tmp_path: Path, *, tire_circumference_m: float = 2.20) -> Path:
+    run_path = tmp_path / "run_content.jsonl"
+    records: list[dict[str, object]] = [_run_metadata(tire_circumference_m=tire_circumference_m)]
+    for idx in range(30):
+        speed = 40 + idx
+        wheel_hz = (speed * KMH_TO_MPS) / tire_circumference_m
+        records.append(
+            _sample(idx, speed_kmh=float(speed), dominant_freq_hz=wheel_hz, peak_amp_g=0.09),
+        )
+    records.append(RUN_END)
+    write_jsonl(run_path, records)
+    return run_path
+
+
+_RECAPTURE_SECTION_HEADING_KEYS = [
+    "REPORT_ACTIONS_PANEL_TITLE",
+    "REPORT_PROOF_PANEL_TITLE_INCONCLUSIVE",
+    "REPORT_RECAPTURE_GUIDANCE_TITLE",
+    "REPORT_CAPTURE_ISSUES_TITLE",
+    "REPORT_CAPTURE_CHANGES_TITLE",
+    "REPORT_CAPTURE_CONDITIONS_TITLE",
+    "REPORT_TRACEABILITY_PANEL_TITLE",
+]
+
+
+@pytest.mark.parametrize(
+    ("lang", "i18n_keys"),
+    [
+        pytest.param("en", _RECAPTURE_SECTION_HEADING_KEYS, id="en_recapture_section_headings"),
+        pytest.param("nl", _RECAPTURE_SECTION_HEADING_KEYS, id="nl_recapture_section_headings"),
+    ],
+)
+def test_pdf_contains_i18n_labels(
+    tmp_path: Path,
+    lang: str,
+    i18n_keys: list[str],
+) -> None:
+    run_path = _make_run_jsonl(tmp_path)
+    summary = summarize_log(run_path, lang=lang)
+    pdf = build_report_pdf(build_report_document(prepare_report_input(summary)))
+    text = extract_pdf_text(pdf)
+    i18n = json.loads(_I18N_JSON.read_text(encoding="utf-8"))
+    missing = []
+    for key in i18n_keys:
+        label = i18n[key][lang]
+        if label not in text:
+            missing.append(f"{key} ({label!r})")
+    assert missing == [], f"Missing {lang} labels in PDF: {missing}"
+
+
+def test_full_report_template_contains_peak_db_column_labels() -> None:
+    i18n = json.loads(_I18N_JSON.read_text(encoding="utf-8"))
+    data = ReportDocument(
+        title="Diagnostic worksheet",
+        lang="en",
+        verdict_page=VerdictPageData(
+            suspected_source="Wheel / Tire",
+            inspect_first="Front-Left",
+            action_status="Action-ready",
+            reason_sentence=(
+                "Wheel / Tire remains the strongest source because the repeated pattern "
+                "stayed strongest near Front-Left."
+            ),
+            dominant_corner="Front-Left",
+            location_confidence="Strong",
+            coverage_label="4 of 4 expected positions stayed connected.",
+            proof_summary=(
+                "Front-Left outranked the next location by 2.1x on "
+                "matched-window linear intensity evidence."
+            ),
+        ),
+        appendix_a=AppendixAData(
+            mode="workflow",
+            primary_source="Wheel / Tire",
+            why_primary_first="Wheel / Tire stayed strongest near Front-Left.",
+            ranked_candidates=[
+                RankedCandidateRow(
+                    source_name="Wheel / Tire",
+                    inspect_first="Front-Left",
+                    path_role="Primary path",
+                    reason="Wheel / Tire stayed strongest near Front-Left.",
+                )
+            ],
+        ),
+        appendix_c=AppendixCData(
+            measurement_rows=[
+                MeasurementRow(
+                    measurement_id="M-01",
+                    source_name="Wheel / Tire",
+                    signal_label="1x wheel order",
+                    peak_db=32.0,
+                    strength_db=24.0,
+                    speed_window="50-60 km/h",
+                    dominant_location="Front-Left",
+                )
+            ],
+            speed_band_summary="Repeated energy stayed strongest in the 50-60 km/h window.",
+        ),
+        traceability_rows=[ReportLabelValueRow(label="Run ID", value="run-1")],
+        next_steps=[NextStep(action="Check wheel balance")],
+    )
+
+    text = extract_pdf_text(build_report_pdf(data))
+
+    assert i18n["REPORT_PEAK_DB_COLUMN"]["en"] in text
+    assert i18n["REPORT_STRENGTH_DB_COLUMN"]["en"] in text
+
+
+def test_pdf_additional_observations_heading_for_transient_findings() -> None:
+    data = ReportDocument(
+        title="Diagnostic worksheet",
+        pattern_evidence=PatternEvidence(),
+        lang="en",
+        appendix_c=AppendixCData(
+            observations=["Transient impact evidence was also seen near Front-Left."]
+        ),
+    )
+
+    pdf = build_report_pdf(data)
+    text = extract_pdf_text(pdf)
+    i18n = json.loads(_I18N_JSON.read_text(encoding="utf-8"))
+    assert i18n["ADDITIONAL_OBSERVATIONS"]["en"] in text
+    assert "(22%)" not in text
+
+
+@pytest.mark.parametrize("lang", ["en", "nl"])
+def test_pdf_workflow_appendix_a_headings_render(lang: str) -> None:
+    i18n = json.loads(_I18N_JSON.read_text(encoding="utf-8"))
+    data = ReportDocument(
+        title="Diagnostic worksheet",
+        lang=lang,
+        appendix_a=AppendixAData(
+            mode="workflow",
+            primary_source="Wheel / Tire",
+            alternative_source="Driveline",
+            why_primary_first="Wheel / Tire stayed strongest near Front-Left.",
+            next_if_clean="Move to the driveline path next and inspect Front-Right.",
+            ranked_candidates=[
+                RankedCandidateRow(
+                    source_name="Wheel / Tire",
+                    inspect_first="Front-Left",
+                    path_role="Primary path",
+                    reason="Wheel / Tire stayed strongest near Front-Left.",
+                )
+            ],
+        ),
+        next_steps=[
+            NextStep(
+                action="Check wheel balance",
+                why="The strongest repeated pattern stayed near Front-Left.",
+                confirm="If confirmed, repeat the run to confirm the reduction.",
+                falsify="If balance is clean, move to the driveline path.",
+            )
+        ],
+    )
+
+    text = extract_pdf_text(build_report_pdf(data))
+
+    assert i18n["REPORT_PRIMARY_VS_ALTERNATIVE_TITLE"][lang] in text
+    assert i18n["REPORT_ACTION_MATRIX_TITLE"][lang] in text
